requirements.txt
requests==2.21.0
bs4==0.0.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
sendgrid==6.0.5
google-api-python-client==1.7.8
retry==0.9.2